introduce cellState which tells if a cell is empty, occupied by snake or food
