+++ src/main/java/org/snakegame/constants/CellState.java
@@ -1,0 +1,7 @@
+package org.snakegame.constants;
+
+public enum CellState {
+    SNAKE,
+    EMPTY,
+    FOOD;
+}
